+++ Game/src/game/LuxPlayerHands.h
@@ -138,7 +138,7 @@ private:
 
 	float mfPosAddMul;
 
-	int mlMaxCamRotations;
+	size_t mlMaxCamRotations;
 	tVector3fList mlstCamRotations;
 
 	eLuxHandsState mHandState;
